LibDSP: Optimize note processing
Previously, a collection of notes (Vector or Array) would be created and
promptly deleted for every sample (at least 44 thousand times per
second!). This is obviously bad for performance. Now, RollNotes are a
HashMap correlating pitches to notes, and Track reuses the data
structure in order to avoid reallocations. That is the reason for
introducing the fast clear_with_capacity to HashMap.